% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics_GQL.R
\name{apply_forward_rates_GQL}
\alias{apply_forward_rates_GQL}
\title{Apply forward rates to unfixed cashflows}
\usage{
apply_forward_rates_GQL(
  cashflows,
  forecast_curve,
  accrual_start_col = "accrual_start_date",
  accrual_end_col = "accrual_end_date",
  day_counter = "Actual365Fixed"
)
}
\arguments{
\item{cashflows}{A cashflow schedule tibble.}

\item{forecast_curve}{A QuantLib yield curve or yield curve handle.}

\item{accrual_start_col}{Accrual start date column name.}

\item{accrual_end_col}{Accrual end date column name.}

\item{day_counter}{Day counter name or QuantLib day counter object.}
}
\value{
A tibble with year_fraction and forward_rate.
}
\description{
Apply forward rates to unfixed cashflows
}
