% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{approx_gradient_GQH}
\alias{approx_gradient_GQH}
\title{Approximate an objective gradient}
\usage{
approx_gradient_GQH(parameters, objective_function, epsilon = 1e-08)
}
\arguments{
\item{parameters}{Parameter vector.}

\item{objective_function}{Objective function.}

\item{epsilon}{Finite-difference step.}
}
\value{
Numeric gradient.
}
\description{
Approximate an objective gradient
}
