% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_GQL.R
\name{as_r_date_GQH}
\alias{as_r_date_GQH}
\title{Convert a date value to R Date}
\usage{
as_r_date_GQH(x)
}
\arguments{
\item{x}{QuantLib Date, R Date, or ISO date string.}
}
\value{
An R Date.
}
\description{
Convert a date value to R Date
}
