% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory_GQL.R
\name{asset_swap_fair_clean_price_GQL}
\alias{asset_swap_fair_clean_price_GQL}
\title{AssetSwap fair clean price}
\usage{
asset_swap_fair_clean_price_GQL(asset_swap)
}
\arguments{
\item{asset_swap}{QuantLib AssetSwap object.}
}
\value{
Numeric fair clean price.
}
\description{
AssetSwap fair clean price
}
