% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory_GQL.R
\name{asset_swap_fair_spread_GQL}
\alias{asset_swap_fair_spread_GQL}
\title{AssetSwap fair spread}
\usage{
asset_swap_fair_spread_GQL(asset_swap)
}
\arguments{
\item{asset_swap}{QuantLib AssetSwap object.}
}
\value{
Numeric fair spread.
}
\description{
AssetSwap fair spread
}
