% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory_GQL.R
\name{asset_swap_set_engine_GQL}
\alias{asset_swap_set_engine_GQL}
\title{Attach a discounting engine to an AssetSwap}
\usage{
asset_swap_set_engine_GQL(asset_swap, discount_curve_handle)
}
\arguments{
\item{asset_swap}{QuantLib AssetSwap object.}

\item{discount_curve_handle}{QuantLib YieldTermStructureHandle.}
}
\value{
The original AssetSwap object, invisibly.
}
\description{
Attach a discounting engine to an AssetSwap
}
