% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory_GQL.R
\name{asset_swap_summary_GQL}
\alias{asset_swap_summary_GQL}
\title{Summarise an AssetSwap}
\usage{
asset_swap_summary_GQL(asset_swap)
}
\arguments{
\item{asset_swap}{QuantLib AssetSwap object.}
}
\value{
A tibble with AssetSwap analytics.
}
\description{
Summarise an AssetSwap
}
