% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{black_calculator_table_GQL}
\alias{black_calculator_table_GQL}
\title{Build a Black calculator metric table}
\usage{
black_calculator_table_GQL(
  payoff,
  forward,
  volatility,
  maturity,
  discount_factor
)
}
\arguments{
\item{payoff}{QuantLib payoff.}

\item{forward}{Forward price.}

\item{volatility}{Volatility.}

\item{maturity}{Time to maturity.}

\item{discount_factor}{Discount factor.}
}
\value{
Metric tibble.
}
\description{
Build a Black calculator metric table
}
