% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics_GQL.R
\name{bond_convexity_GQL}
\alias{bond_convexity_GQL}
\title{Bond Convexity}
\usage{
bond_convexity_GQL(
  bond,
  ytm = NULL,
  day_counter = QuantLib::Actual365Fixed(),
  compounding = "Compounded",
  frequency = "Annual"
)
}
\description{
Bond Convexity
}
