% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics_GQL.R
\name{bond_coupon_info_GQL}
\alias{bond_coupon_info_GQL}
\title{Get bond coupon information}
\usage{
bond_coupon_info_GQL(bond, as_of = eval_date_get_GQL())
}
\arguments{
\item{bond}{A QuantLib bond object.}

\item{as_of}{Evaluation date. Default is  eval_date_get_GQL().}
}
\value{
A tibble with coupon cashflow information.
}
\description{
Get bond coupon information
}
