% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics_GQL.R
\name{bond_duration_GQL}
\alias{bond_duration_GQL}
\title{Bond Duration}
\usage{
bond_duration_GQL(
  bond,
  ytm = NULL,
  day_counter = QuantLib::Actual365Fixed(),
  compounding = "Compounded",
  frequency = "Annual"
)
}
\description{
Bond Duration
}
