% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics_GQL.R
\name{bond_dv01_GQL}
\alias{bond_dv01_GQL}
\title{Bond DV01}
\usage{
bond_dv01_GQL(
  bond,
  ytm = NULL,
  bp = 1e-04,
  day_counter = QuantLib::Actual360(),
  compounding = "Compounded",
  frequency = "Annual"
)
}
\description{
Dollar value of 1bp yield change.
}
