% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics_GQL.R
\name{bond_flat_forward_handle_GQL}
\alias{bond_flat_forward_handle_GQL}
\title{Build a flat-forward yield-curve handle for bond pricing}
\usage{
bond_flat_forward_handle_GQL(
  settlement_date,
  rate,
  day_counter = QuantLib::Actual360(),
  compounding = "Compounded",
  frequency = "Annual"
)
}
\arguments{
\item{settlement_date}{Settlement date.}

\item{rate}{Flat rate.}

\item{day_counter}{QuantLib day counter.}

\item{compounding}{QuantLib compounding convention or supported string.}

\item{frequency}{QuantLib frequency convention or supported string.}
}
\value{
QuantLib YieldTermStructureHandle.
}
\description{
Build a flat-forward yield-curve handle for bond pricing
}
