% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics_GQL.R
\name{bond_npv_with_curve_GQL}
\alias{bond_npv_with_curve_GQL}
\title{Bond NPV with a curve handle}
\usage{
bond_npv_with_curve_GQL(bond, curve_handle)
}
\arguments{
\item{bond}{QuantLib bond object.}

\item{curve_handle}{QuantLib YieldTermStructureHandle.}
}
\value{
Numeric NPV.
}
\description{
Bond NPV with a curve handle
}
