% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics_GQL.R
\name{bond_npv_with_flat_yield_GQL}
\alias{bond_npv_with_flat_yield_GQL}
\title{Bond NPV with a flat yield}
\usage{
bond_npv_with_flat_yield_GQL(
  bond,
  settlement_date,
  rate,
  day_counter = QuantLib::Actual360(),
  compounding = "Compounded",
  frequency = "Annual"
)
}
\arguments{
\item{bond}{QuantLib bond object.}

\item{settlement_date}{Settlement date.}

\item{rate}{Flat rate.}

\item{day_counter}{QuantLib day counter.}

\item{compounding}{QuantLib compounding convention or supported string.}

\item{frequency}{QuantLib frequency convention or supported string.}
}
\value{
Numeric NPV.
}
\description{
Bond NPV with a flat yield
}
