% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics_GQL.R
\name{bond_npv_with_zspread_GQL}
\alias{bond_npv_with_zspread_GQL}
\title{Bond NPV with a z-spreaded curve}
\usage{
bond_npv_with_zspread_GQL(
  bond,
  base_curve_handle,
  z_spread,
  compounding = "Compounded",
  frequency = "Annual",
  day_counter = QuantLib::Actual360()
)
}
\arguments{
\item{bond}{QuantLib bond object.}

\item{base_curve_handle}{Base QuantLib YieldTermStructureHandle.}

\item{z_spread}{Z-spread.}

\item{compounding}{QuantLib compounding convention or supported string.}

\item{frequency}{QuantLib frequency convention or supported string.}

\item{day_counter}{QuantLib day counter.}
}
\value{
A list with spread curve, spread curve handle, and NPV.
}
\description{
Bond NPV with a z-spreaded curve
}
