% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_yen_GQL.R
\name{bond_price_from_simple_rate_YEN_GQL}
\alias{bond_price_from_simple_rate_YEN_GQL}
\title{Price a yen-denominated bond from a simple rate}
\usage{
bond_price_from_simple_rate_YEN_GQL(
  start_date,
  end_date,
  coupon_rate,
  simple_rate,
  redemption = 100,
  digits = 3L
)
}
\arguments{
\item{start_date}{Settlement date.}

\item{end_date}{Maturity date.}

\item{coupon_rate}{Annual coupon amount per 100 face value, in percent.}

\item{simple_rate}{Annual simple rate, in percent.}

\item{redemption}{Redemption amount per 100 face value.}

\item{digits}{Optional number of decimal places to truncate to. Use \code{NULL}
for the untruncated result.}
}
\value{
Bond price per 100 face value.
}
\description{
Price a yen-denominated bond from a simple rate
}
