% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics_GQL.R
\name{bond_price_from_yield_GQL}
\alias{bond_price_from_yield_GQL}
\title{Bond Price from Yield}
\usage{
bond_price_from_yield_GQL(
  bond,
  ytm,
  day_counter = QuantLib::Actual360(),
  compounding = "Compounded",
  frequency = "Annual",
  settlement_date = NULL
)
}
\description{
Bond Price from Yield
}
