% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics_GQL.R
\name{bond_price_yield_curve_GQL}
\alias{bond_price_yield_curve_GQL}
\title{Bond Price-Yield Curve}
\usage{
bond_price_yield_curve_GQL(
  bond,
  ytm_center = NULL,
  spread = 0.02,
  n = 41,
  day_counter = QuantLib::Actual360(),
  compounding = "Compounded",
  frequency = "Annual"
)
}
\description{
Bond Price-Yield Curve
}
