% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics_GQL.R
\name{bond_risk_measures_GQL}
\alias{bond_risk_measures_GQL}
\title{Bond risk measures from yield}
\usage{
bond_risk_measures_GQL(
  bond,
  yield,
  day_counter = QuantLib::Actual360(),
  compounding = "Compounded",
  frequency = "Annual"
)
}
\arguments{
\item{bond}{QuantLib bond object.}

\item{yield}{Yield to maturity.}

\item{day_counter}{QuantLib day counter.}

\item{compounding}{QuantLib compounding convention or supported string.}

\item{frequency}{QuantLib frequency convention or supported string.}
}
\value{
A tibble with duration, BPV/PV01, and convexity measures.
}
\description{
Bond risk measures from yield
}
