% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics_GQL.R
\name{bond_settlement_info_GQL}
\alias{bond_settlement_info_GQL}
\title{Get bond settlement information}
\usage{
bond_settlement_info_GQL(bond, as_of = eval_date_get_GQL())
}
\description{
Get bond settlement information
}
