% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{bond_yield_from_clean_safe_GQL}
\alias{bond_yield_from_clean_safe_GQL}
\title{Safely calculate bond yield from clean price}
\usage{
bond_yield_from_clean_safe_GQL(
  bond,
  clean_price,
  day_counter,
  compounding,
  frequency
)
}
\arguments{
\item{bond}{QuantLib bond.}

\item{clean_price}{Clean price.}

\item{day_counter}{QuantLib day counter.}

\item{compounding}{Compounding convention.}

\item{frequency}{Coupon frequency.}
}
\value{
Numeric yield.
}
\description{
Safely calculate bond yield from clean price
}
