% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics_GQL.R
\name{bond_yield_from_price_GQL}
\alias{bond_yield_from_price_GQL}
\title{Bond Yield from Clean Price}
\usage{
bond_yield_from_price_GQL(
  bond,
  clean_price,
  day_counter = QuantLib::Actual360(),
  compounding = "Compounded",
  frequency = "Annual",
  settlement_date = NULL
)
}
\description{
Bond Yield from Clean Price
}
