% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics_GQL.R
\name{bond_yield_GQL}
\alias{bond_yield_GQL}
\title{Bond Yield}
\usage{
bond_yield_GQL(
  bond,
  day_counter = QuantLib::Actual365Fixed(),
  compounding = "Compounded",
  frequency = "Annual"
)
}
\description{
Bond Yield
}
