% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_builder_GQL.R
\name{build_discount_curve_GQL}
\alias{build_discount_curve_GQL}
\title{Build a QuantLib discount curve from zero-rate nodes}
\usage{
build_discount_curve_GQL(nodes, day_counter = "Actual365Fixed")
}
\arguments{
\item{nodes}{A data frame with date and zero_rate columns.}

\item{day_counter}{Day counter name or QuantLib day counter object.}
}
\value{
A QuantLib DiscountCurve object.
}
\description{
Build a QuantLib discount curve from zero-rate nodes
}
