% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory_GQL.R
\name{build_eonia_ois_GQL}
\alias{build_eonia_ois_GQL}
\title{Build an Eonia OIS swap}
\usage{
build_eonia_ois_GQL(swap_tenor, forecast_handle, fixed_rate)
}
\arguments{
\item{swap_tenor}{QuantLib Period object.}

\item{forecast_handle}{QuantLib yield term structure handle.}

\item{fixed_rate}{Fixed rate.}
}
\value{
A QuantLib OIS swap object.
}
\description{
Build an Eonia OIS swap
}
