% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory_GQL.R
\name{build_ois_from_trade_GQL}
\alias{build_ois_from_trade_GQL}
\title{Build an OIS from trade data}
\usage{
build_ois_from_trade_GQL(trade, forecast_handle)
}
\arguments{
\item{trade}{A one-row data frame containing OIS trade fields.}

\item{forecast_handle}{QuantLib forecast curve handle.}
}
\value{
A QuantLib OIS object.
}
\description{
Build an OIS from trade data
}
