% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory_GQL.R
\name{build_ois_GQL}
\alias{build_ois_GQL}
\title{Build an OIS swap}
\usage{
build_ois_GQL(swap_tenor, overnight_index, fixed_rate)
}
\arguments{
\item{swap_tenor}{QuantLib Period object.}

\item{overnight_index}{QuantLib overnight index object.}

\item{fixed_rate}{Fixed rate.}
}
\value{
A QuantLib OIS swap object.
}
\description{
Build an OIS swap
}
