% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_builder_GQL.R
\name{build_swap_curve_GQL}
\alias{build_swap_curve_GQL}
\title{Build a deposit-swap forecasting curve}
\usage{
build_swap_curve_GQL(settlement_date = "2008-09-18", fixing_days = 3)
}
\arguments{
\item{settlement_date}{Settlement date.}

\item{fixing_days}{Fixing days for deposit helpers.}
}
\value{
A QuantLib yield term structure.
}
\description{
Build a deposit-swap forecasting curve
}
