% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{callable_bond_dirty_price_safe_GQL}
\alias{callable_bond_dirty_price_safe_GQL}
\title{Safely obtain a callable bond dirty price}
\usage{
callable_bond_dirty_price_safe_GQL(bond)
}
\arguments{
\item{bond}{QuantLib callable bond.}
}
\value{
Numeric price or NA.
}
\description{
Safely obtain a callable bond dirty price
}
