% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option_GQL.R
\name{cap_floor_optionlet_prices_GQL}
\alias{cap_floor_optionlet_prices_GQL}
\title{Extract cap/floor optionlet prices}
\usage{
cap_floor_optionlet_prices_GQL(cap_floor)
}
\description{
Extract cap/floor optionlet prices
}
