% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics_GQL.R
\name{cashflow_leg_summary_GQL}
\alias{cashflow_leg_summary_GQL}
\title{Summarise cashflows by leg}
\usage{
cashflow_leg_summary_GQL(cashflows)
}
\arguments{
\item{cashflows}{A cashflow schedule tibble.}
}
\value{
A tibble summarising cashflows by leg.
}
\description{
Summarise cashflows by leg
}
