% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics_GQL.R
\name{cashflow_schedule_from_trade_GQL}
\alias{cashflow_schedule_from_trade_GQL}
\title{Build cashflow schedule from one trade}
\usage{
cashflow_schedule_from_trade_GQL(trade, forecast_handle)
}
\arguments{
\item{trade}{A one-row trade data frame.}

\item{forecast_handle}{Forecast curve handle.}
}
\value{
A tibble containing cashflows.
}
\description{
Build cashflow schedule from one trade
}
