% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics_GQL.R
\name{cashflow_schedule_from_trades_GQL}
\alias{cashflow_schedule_from_trades_GQL}
\title{Build cashflow schedules from multiple trades}
\usage{
cashflow_schedule_from_trades_GQL(trades, forecast_handle)
}
\arguments{
\item{trades}{A trade data frame.}

\item{forecast_handle}{Forecast curve handle.}
}
\value{
A tibble containing cashflows for all supported trades.
}
\description{
Build cashflow schedules from multiple trades
}
