% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{cds_buyer_side_GQL}
\alias{cds_buyer_side_GQL}
\title{Obtain the QuantLib CDS buyer-side enum}
\usage{
cds_buyer_side_GQL()
}
\value{
QuantLib protection buyer enum.
}
\description{
Obtain the QuantLib CDS buyer-side enum
}
