% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory_GQL.R
\name{cds_fair_spread_GQL}
\alias{cds_fair_spread_GQL}
\title{CDS fair spread}
\usage{
cds_fair_spread_GQL(cds)
}
\arguments{
\item{cds}{QuantLib CreditDefaultSwap object.}
}
\value{
Numeric fair spread.
}
\description{
CDS fair spread
}
