% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{cds_isda_engine_GQL}
\alias{cds_isda_engine_GQL}
\title{Build an ISDA CDS pricing engine}
\usage{
cds_isda_engine_GQL(hazard_curve_handle, recovery_rate, discount_curve_handle)
}
\arguments{
\item{hazard_curve_handle}{Default probability curve handle.}

\item{recovery_rate}{Recovery rate.}

\item{discount_curve_handle}{Discount curve handle.}
}
\value{
QuantLib IsdaCdsEngine.
}
\description{
Build an ISDA CDS pricing engine
}
