% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory_GQL.R
\name{cds_summary_GQL}
\alias{cds_summary_GQL}
\title{Summarise a CDS}
\usage{
cds_summary_GQL(cds)
}
\arguments{
\item{cds}{QuantLib CreditDefaultSwap object.}
}
\value{
A tibble with CDS analytics.
}
\description{
Summarise a CDS
}
