% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_period_utils_GQL.R
\name{chr_GQL}
\alias{chr_GQL}
\title{Convert QuantLib object to character}
\usage{
chr_GQL(x)
}
\arguments{
\item{x}{Object to print.}
}
\value{
Character scalar.
}
\description{
Convert QuantLib object to character
}
