% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{curve_discount_safe_GQL}
\alias{curve_discount_safe_GQL}
\title{Safely obtain a discount factor}
\usage{
curve_discount_safe_GQL(curve, x)
}
\arguments{
\item{curve}{QuantLib yield curve.}

\item{x}{QuantLib Date or time.}
}
\value{
Numeric discount factor or NA.
}
\description{
Safely obtain a discount factor
}
