% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_helpers_GQH.R
\name{curve_table_GQH}
\alias{curve_table_GQH}
\title{Alias for curve_tbl_GQH}
\usage{
curve_table_GQH(
  curve,
  tenors = c("1D", "1W", "1M", "3M", "6M", "1Y", "2Y", "3Y", "5Y", "7Y", "10Y", "20Y",
    "30Y")
)
}
\description{
Alias for curve_tbl_GQH
}
