% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_GQL.R
\name{date_GQL}
\alias{date_GQL}
\title{Create a QuantLib Date}
\usage{
date_GQL(x)
}
\arguments{
\item{x}{A character scalar in ISO date format or an R \code{Date}.}
}
\value{
A QuantLib Date object.
}
\description{
Create a QuantLib Date
}
