% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_GQL.R
\name{DateParser_parseISO_GQL}
\alias{DateParser_parseISO_GQL}
\title{Parse an ISO date string as a QuantLib Date}
\usage{
DateParser_parseISO_GQL(x)
}
\arguments{
\item{x}{A character scalar in ISO date format or an R \code{Date}.}
}
\value{
A QuantLib Date object.
}
\description{
Converts a character scalar in \code{YYYY-MM-DD} format, or an R \code{Date},
to a QuantLib Date object.
}
