% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_GQL.R
\name{day_counter_GQL}
\alias{day_counter_GQL}
\title{Select QuantLib day counter}
\usage{
day_counter_GQL(day_counter = "Actual365Fixed")
}
\arguments{
\item{day_counter}{Day counter name or QuantLib day counter object.}
}
\value{
QuantLib day counter object.
}
\description{
Select QuantLib day counter
}
