% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_bootstrap_estr_GQL.R
\name{estr_GQL}
\alias{estr_GQL}
\title{Create an ESTR overnight index}
\usage{
estr_GQL(forwarding_curve_handle = NULL)
}
\arguments{
\item{forwarding_curve_handle}{Optional QuantLib yield-term-structure
handle.}
}
\value{
A QuantLib ESTR overnight-index object.
}
\description{
Create an ESTR overnight index
}
