% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_GQL.R
\name{eval_date_GQL}
\alias{eval_date_GQL}
\title{Set QuantLib evaluation date}
\usage{
eval_date_GQL(x)
}
\arguments{
\item{x}{Date or character scalar.}
}
\value{
Invisibly returns QuantLib Date object.
}
\description{
Set QuantLib evaluation date
}
