% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_builder_GQL.R
\name{example_ois_nodes_GQL}
\alias{example_ois_nodes_GQL}
\title{Example OIS curve nodes}
\usage{
example_ois_nodes_GQL()
}
\value{
A tibble of sample zero-rate nodes.
}
\description{
Example OIS curve nodes
}
