% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{flat_curve_GQL}
\alias{flat_curve_GQL}
\title{Build a flat yield curve}
\usage{
flat_curve_GQL(
  reference_date,
  rate,
  day_counter = QuantLib::Actual365Fixed(),
  compounding = "Continuous"
)
}
\arguments{
\item{reference_date}{QuantLib Date.}

\item{rate}{Flat rate.}

\item{day_counter}{QuantLib day counter.}

\item{compounding}{Compounding convention.}
}
\value{
QuantLib FlatForward object.
}
\description{
Build a flat yield curve
}
