% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_helpers_GQH.R
\name{fmt_num_GQH}
\alias{fmt_num_GQH}
\title{Format a numeric value for display}
\usage{
fmt_num_GQH(x, digits = 6)
}
\arguments{
\item{x}{Numeric vector.}

\item{digits}{Number of digits.}
}
\value{
Character vector.
}
\description{
Format a numeric value for display
}
