% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{forward_swap_rate_from_schedule_GQL}
\alias{forward_swap_rate_from_schedule_GQL}
\title{Calculate a forward swap rate from a schedule}
\usage{
forward_swap_rate_from_schedule_GQL(schedule, curve, day_counter)
}
\arguments{
\item{schedule}{QuantLib Schedule.}

\item{curve}{QuantLib yield curve.}

\item{day_counter}{QuantLib day counter.}
}
\value{
Numeric forward swap rate.
}
\description{
Calculate a forward swap rate from a schedule
}
