% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exotic_derivatice_FX_IR.R
\name{fx_process_GQL}
\alias{fx_process_GQL}
\title{Build a Garman-Kohlhagen FX process}
\usage{
fx_process_GQL(
  spot,
  foreign_curve_handle,
  domestic_curve_handle,
  volatility_handle
)
}
\arguments{
\item{spot}{Numeric FX spot or QuantLib QuoteHandle.}

\item{foreign_curve_handle}{QuantLib foreign yield-curve handle.}

\item{domestic_curve_handle}{QuantLib domestic yield-curve handle.}

\item{volatility_handle}{QuantLib Black volatility handle.}
}
\value{
QuantLib GarmanKohlagenProcess.
}
\description{
Build a Garman-Kohlhagen FX process
}
