% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{generate_path_table_GQL}
\alias{generate_path_table_GQL}
\title{Generate a path table}
\usage{
generate_path_table_GQL(
  process,
  maturity,
  n_steps,
  n_paths,
  seed = 1L,
  sequence = c("pseudo", "sobol")
)
}
\arguments{
\item{process}{QuantLib stochastic process.}

\item{maturity}{Path maturity.}

\item{n_steps}{Number of steps.}

\item{n_paths}{Number of paths.}

\item{seed}{Random seed.}

\item{sequence}{Pseudo-random or Sobol.}
}
\value{
Long-form path tibble.
}
\description{
Generate a path table
}
