% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{get_metric_num_GQH}
\alias{get_metric_num_GQH}
\title{Extract a numeric metric value}
\usage{
get_metric_num_GQH(tbl, metric_name)
}
\arguments{
\item{tbl}{Metric table.}

\item{metric_name}{Metric name.}
}
\value{
Numeric scalar or NA.
}
\description{
Extract a numeric metric value
}
