% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{hull_white_a_GQL}
\alias{hull_white_a_GQL}
\title{Calculate Hull-White A}
\usage{
hull_white_a_GQL(time, maturity, mean_reversion, sigma)
}
\arguments{
\item{time}{Start time.}

\item{maturity}{Maturity.}

\item{mean_reversion}{Mean reversion.}

\item{sigma}{Short-rate volatility.}
}
\value{
Numeric A.
}
\description{
Calculate Hull-White A
}
