% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{hull_white_b_GQL}
\alias{hull_white_b_GQL}
\title{Calculate Hull-White B}
\usage{
hull_white_b_GQL(time, maturity, mean_reversion)
}
\arguments{
\item{time}{Start time.}

\item{maturity}{Maturity.}

\item{mean_reversion}{Mean reversion.}
}
\value{
Numeric B.
}
\description{
Calculate Hull-White B
}
