% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{hull_white_forward_drift_GQL}
\alias{hull_white_forward_drift_GQL}
\title{Calculate an approximate Hull-White forward-measure drift}
\usage{
hull_white_forward_drift_GQL(
  start_time,
  exercise_time,
  maturity,
  mean_reversion,
  sigma
)
}
\arguments{
\item{start_time}{Start time.}

\item{exercise_time}{Exercise time.}

\item{maturity}{Maturity.}

\item{mean_reversion}{Mean reversion.}

\item{sigma}{Short-rate volatility.}
}
\value{
Numeric drift adjustment.
}
\description{
Calculate an approximate Hull-White forward-measure drift
}
