% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option_GQL.R
\name{hull_white_swaption_engine_GQL}
\alias{hull_white_swaption_engine_GQL}
\title{Make a Hull-White swaption pricing engine}
\usage{
hull_white_swaption_engine_GQL(
  term_structure = NULL,
  valuation_date = eval_date_get_GQL(),
  rate = 0.03,
  a = 0.03,
  sigma = 0.01,
  method = c("jamshidian", "tree"),
  time_steps = 60L,
  day_counter = QuantLib::Actual365Fixed()
)
}
\arguments{
\item{term_structure}{Optional QuantLib yield term structure handle.}

\item{valuation_date}{Evaluation date used when term_structure is NULL.}

\item{rate}{Flat rate used when term_structure is NULL.}

\item{a}{Hull-White mean reversion.}

\item{sigma}{Hull-White volatility.}

\item{method}{Pricing method. Use "jamshidian" or "tree".}

\item{time_steps}{Number of tree time steps when method = "tree".}

\item{day_counter}{QuantLib day counter.}
}
\value{
QuantLib swaption pricing engine.
}
\description{
Make a Hull-White swaption pricing engine
}
