% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{instrument_npv_safe_GQL}
\alias{instrument_npv_safe_GQL}
\title{Safely calculate an instrument NPV}
\usage{
instrument_npv_safe_GQL(instrument)
}
\arguments{
\item{instrument}{QuantLib instrument.}
}
\value{
Numeric NPV or NA.
}
\description{
Safely calculate an instrument NPV
}
