% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat_quantive_riemann_GQH.R
\name{ir_basis_table_GQH}
\alias{ir_basis_table_GQH}
\title{Build a basis curve table}
\usage{
ir_basis_table_GQH(
  basis_env,
  tenors = c("1M", "3M", "6M", "1Y", "2Y", "3Y", "5Y", "10Y", "20Y", "30Y")
)
}
\description{
Build a basis curve table
}
