% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat_quantive_riemann_GQH.R
\name{ir_make_basis_curve_GQH}
\alias{ir_make_basis_curve_GQH}
\title{Make a simple interpolated basis curve environment}
\usage{
ir_make_basis_curve_GQH(
  basis_data,
  base_curve_env,
  spread_label = "basis",
  verbose = TRUE
)
}
\description{
Make a simple interpolated basis curve environment
}
