% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_GQL.R
\name{iso_GQL}
\alias{iso_GQL}
\title{Convert QuantLib Date to ISO string}
\usage{
iso_GQL(x)
}
\arguments{
\item{x}{QuantLib Date object.}
}
\value{
ISO date string.
}
\description{
Convert QuantLib Date to ISO string
}
