% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics_GQL.R
\name{leg_cashflow_schedule_GQL}
\alias{leg_cashflow_schedule_GQL}
\title{Build cashflow schedule table from a QuantLib leg}
\usage{
leg_cashflow_schedule_GQL(leg, leg_name, leg_no = NA_integer_)
}
\arguments{
\item{leg}{A QuantLib Leg object.}

\item{leg_name}{Leg name.}

\item{leg_no}{Leg number. QuantLib uses 0-based leg indexing.}
}
\value{
A tibble containing cashflows for one leg.
}
\description{
Build cashflow schedule table from a QuantLib leg
}
