% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_GQL.R
\name{leg_to_cashflow_tbl_GQL}
\alias{leg_to_cashflow_tbl_GQL}
\title{Convert QuantLib leg to cashflow table}
\usage{
leg_to_cashflow_tbl_GQL(leg)
}
\arguments{
\item{leg}{QuantLib Leg or CashFlow vector.}
}
\value{
A tibble with cashflow dates and amounts.
}
\description{
Convert QuantLib leg to cashflow table
}
