% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_analytics_GQL.R
\name{make_barrier_option_from_trade_GQL}
\alias{make_barrier_option_from_trade_GQL}
\title{Make a barrier option from trade data}
\usage{
make_barrier_option_from_trade_GQL(trade, pricing_engine = NULL)
}
\arguments{
\item{trade}{A one-row data frame containing barrier option trade fields.}

\item{pricing_engine}{Optional QuantLib pricing engine.}
}
\value{
QuantLib BarrierOption object.
}
\description{
Make a barrier option from trade data
}
