% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option_GQL.R
\name{make_bermudan_swaption_from_trade_GQL}
\alias{make_bermudan_swaption_from_trade_GQL}
\title{Make a Bermudan swaption from trade data}
\usage{
make_bermudan_swaption_from_trade_GQL(
  trade,
  forecast_handle = NULL,
  discount_handle = NULL,
  pricing_engine = NULL
)
}
\arguments{
\item{trade}{A one-row data frame containing Bermudan swaption trade fields.}

\item{forecast_handle}{Optional forecast curve handle.}

\item{discount_handle}{Optional discount curve handle.}

\item{pricing_engine}{Optional QuantLib swaption pricing engine.}
}
\value{
QuantLib Swaption object.
}
\description{
Make a Bermudan swaption from trade data
}
