% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option_GQL.R
\name{make_bermudan_swaption_GQL}
\alias{make_bermudan_swaption_GQL}
\title{Make a QuantLib Bermudan swaption}
\usage{
make_bermudan_swaption_GQL(
  underlying_swap,
  exercise_dates,
  pricing_engine = NULL
)
}
\arguments{
\item{underlying_swap}{QuantLib VanillaSwap object.}

\item{exercise_dates}{Bermudan exercise dates.}

\item{pricing_engine}{Optional QuantLib pricing engine.}
}
\value{
QuantLib Swaption object.
}
\description{
Make a QuantLib Bermudan swaption
}
