% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{make_callability_schedule_GQL}
\alias{make_callability_schedule_GQL}
\title{Build a QuantLib callability schedule}
\usage{
make_callability_schedule_GQL(call_dates, call_price_clean = 100)
}
\arguments{
\item{call_dates}{Character, Date, or QuantLib dates.}

\item{call_price_clean}{Clean call price.}
}
\value{
QuantLib CallabilitySchedule.
}
\description{
Build a QuantLib callability schedule
}
