% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option_GQL.R
\name{make_cap_from_trade_GQL}
\alias{make_cap_from_trade_GQL}
\title{Make a cap from trade data}
\usage{
make_cap_from_trade_GQL(trade, pricing_engine = NULL)
}
\arguments{
\item{trade}{A one-row data frame containing cap trade fields.}

\item{pricing_engine}{Optional QuantLib pricing engine.}
}
\value{
QuantLib Cap object.
}
\description{
Make a cap from trade data
}
