% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_GQL.R
\name{make_date_vector_GQL}
\alias{make_date_vector_GQL}
\title{Build QuantLib DateVector}
\usage{
make_date_vector_GQL(dates)
}
\arguments{
\item{dates}{Character vector or Date vector.}
}
\value{
QuantLib DateVector.
}
\description{
Build QuantLib DateVector
}
