% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option_GQL.R
\name{make_floor_from_trade_GQL}
\alias{make_floor_from_trade_GQL}
\title{Make a floor from trade data}
\usage{
make_floor_from_trade_GQL(trade, pricing_engine = NULL)
}
\arguments{
\item{trade}{A one-row data frame containing floor trade fields.}

\item{pricing_engine}{Optional QuantLib pricing engine.}
}
\value{
QuantLib Floor object.
}
\description{
Make a floor from trade data
}
