% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option_GQL.R
\name{make_floor_GQL}
\alias{make_floor_GQL}
\title{Make a QuantLib interest-rate floor}
\usage{
make_floor_GQL(
  notional,
  start_date,
  maturity_date,
  floor_rate,
  valuation_date = eval_date_get_GQL(),
  tenor_months = 6L,
  discount_rate = 0.03,
  forecast_rate = discount_rate,
  volatility = 0.2,
  day_counter = QuantLib::Actual365Fixed(),
  calendar = QuantLib::TARGET(),
  convention = QuantLib::BusinessDayConvention_ModifiedFollowing_get(),
  pricing_engine = NULL
)
}
\description{
Make a QuantLib interest-rate floor
}
