% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory_GQL.R
\name{make_ois_from_trade_GQL}
\alias{make_ois_from_trade_GQL}
\title{Make an OIS from trade data}
\usage{
make_ois_from_trade_GQL(trade, forecast_handle)
}
\arguments{
\item{trade}{A one-row data frame containing OIS trade fields.}

\item{forecast_handle}{QuantLib forecast curve handle.}
}
\value{
A QuantLib OIS object.
}
\description{
Make an OIS from trade data
}
