% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory_GQL.R
\name{make_vanilla_swap_from_trade_GQL}
\alias{make_vanilla_swap_from_trade_GQL}
\title{Make a QuantLib VanillaSwap from trade data}
\usage{
make_vanilla_swap_from_trade_GQL(
  trade,
  forecast_handle = NULL,
  discount_handle = NULL
)
}
\arguments{
\item{trade}{A one-row data frame containing VanillaSwap trade fields.}

\item{forecast_handle}{Forecast curve handle.}

\item{discount_handle}{Discount curve handle. If supplied, a DiscountingSwapEngine is attached.}
}
\value{
A QuantLib VanillaSwap object.
}
\description{
Make a QuantLib VanillaSwap from trade data
}
