% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{metric_tbl_display_GQH}
\alias{metric_tbl_display_GQH}
\title{Format a metric table for display}
\usage{
metric_tbl_display_GQH(tbl)
}
\arguments{
\item{tbl}{Metric table.}
}
\value{
A tibble with a character value column when present.
}
\description{
Format a metric table for display
}
