% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{metric_value_to_chr_GQH}
\alias{metric_value_to_chr_GQH}
\title{Convert a metric value to character}
\usage{
metric_value_to_chr_GQH(x)
}
\arguments{
\item{x}{Value or list-column element.}
}
\value{
Character scalar.
}
\description{
Convert a metric value to character
}
