% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{normal_option_greeks_GQL}
\alias{normal_option_greeks_GQL}
\title{Calculate normal-model option Greeks}
\usage{
normal_option_greeks_GQL(
  option_sign,
  forward,
  strike,
  vol,
  maturity,
  discount_factor,
  risk_free_rate
)
}
\arguments{
\item{option_sign}{1 for call, -1 for put.}

\item{forward}{Forward.}

\item{strike}{Strike.}

\item{vol}{Normal volatility.}

\item{maturity}{Maturity.}

\item{discount_factor}{Discount factor.}

\item{risk_free_rate}{Risk-free rate.}
}
\value{
Metric tibble.
}
\description{
Calculate normal-model option Greeks
}
