% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{normal_option_price_GQL}
\alias{normal_option_price_GQL}
\title{Price an option under the normal model}
\usage{
normal_option_price_GQL(
  option_sign,
  forward,
  strike,
  vol,
  maturity,
  discount_factor
)
}
\arguments{
\item{option_sign}{1 for call, -1 for put.}

\item{forward}{Forward.}

\item{strike}{Strike.}

\item{vol}{Normal volatility.}

\item{maturity}{Maturity.}

\item{discount_factor}{Discount factor.}
}
\value{
Numeric option value.
}
\description{
Price an option under the normal model
}
