% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{normal_vol_hagan_GQL}
\alias{normal_vol_hagan_GQL}
\title{Calculate Hagan normal SABR volatility}
\usage{
normal_vol_hagan_GQL(strike, forward, maturity, beta, alpha, volvol, rho)
}
\arguments{
\item{strike}{Strike.}

\item{forward}{Forward.}

\item{maturity}{Maturity.}

\item{beta}{SABR beta.}

\item{alpha}{SABR alpha.}

\item{volvol}{SABR volatility of volatility.}

\item{rho}{SABR correlation.}
}
\value{
Numeric normal volatility.
}
\description{
Calculate Hagan normal SABR volatility
}
