% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics_GQL.R
\name{ois_cashflow_schedule_GQL}
\alias{ois_cashflow_schedule_GQL}
\title{Build OIS cashflow schedule table from fixed and overnight legs}
\usage{
ois_cashflow_schedule_GQL(swap)
}
\arguments{
\item{swap}{A QuantLib OIS-like object.}
}
\value{
A tibble containing fixed and overnight leg cashflows.
}
\description{
Build OIS cashflow schedule table from fixed and overnight legs
}
