% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_builder_GQL.R
\name{ois_curve_GQL}
\alias{ois_curve_GQL}
\title{Build a tidy OIS-style discount curve from zero-rate nodes}
\usage{
ois_curve_GQL(nodes_tbl, day_counter = "Actual365Fixed")
}
\arguments{
\item{nodes_tbl}{A data frame with date and zero_rate columns.}

\item{day_counter}{Day counter name or QuantLib day counter object.}
}
\value{
A list with curve and table.
}
\description{
Build a tidy OIS-style discount curve from zero-rate nodes
}
