% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics_GQL.R
\name{ois_overnight_leg_table_GQL}
\alias{ois_overnight_leg_table_GQL}
\title{OIS overnight leg cashflow table}
\usage{
ois_overnight_leg_table_GQL(swap)
}
\arguments{
\item{swap}{A QuantLib OIS-like object.}
}
\value{
A tibble of overnight leg cashflows.
}
\description{
OIS overnight leg cashflow table
}
