% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics_GQL.R
\name{ois_summary_GQL}
\alias{ois_summary_GQL}
\title{Summarise OIS cashflow legs}
\usage{
ois_summary_GQL(fixed_leg, overnight_leg)
}
\arguments{
\item{fixed_leg}{Fixed leg cashflow table.}

\item{overnight_leg}{Overnight leg cashflow table.}
}
\value{
A tibble summarising OIS cashflow legs.
}
\description{
Summarise OIS cashflow legs
}
