% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_analytics_GQL.R
\name{option_delta_GQL}
\alias{option_delta_GQL}
\title{Option delta}
\usage{
option_delta_GQL(option)
}
\arguments{
\item{option}{QuantLib option object.}
}
\value{
Numeric delta.
}
\description{
Option delta
}
