% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_analytics_GQL.R
\name{option_gamma_GQL}
\alias{option_gamma_GQL}
\title{Option gamma}
\usage{
option_gamma_GQL(option)
}
\arguments{
\item{option}{QuantLib option object.}
}
\value{
Numeric gamma.
}
\description{
Option gamma
}
