% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{option_metric_table_GQL}
\alias{option_metric_table_GQL}
\title{Build an option metric table}
\usage{
option_metric_table_GQL(option, process = NULL)
}
\arguments{
\item{option}{QuantLib option.}

\item{process}{Optional pricing process.}
}
\value{
Metric tibble.
}
\description{
Build an option metric table
}
