% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_analytics_GQL.R
\name{option_npv_GQL}
\alias{option_npv_GQL}
\title{Option NPV}
\usage{
option_npv_GQL(option)
}
\arguments{
\item{option}{QuantLib option object.}
}
\value{
Numeric NPV.
}
\description{
Option NPV
}
