% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_analytics_GQL.R
\name{option_rho_GQL}
\alias{option_rho_GQL}
\title{Option rho}
\usage{
option_rho_GQL(option)
}
\arguments{
\item{option}{QuantLib option object.}
}
\value{
Numeric rho.
}
\description{
Option rho
}
