% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_analytics_GQL.R
\name{option_summary_GQL}
\alias{option_summary_GQL}
\title{Summarise an option}
\usage{
option_summary_GQL(option)
}
\arguments{
\item{option}{QuantLib option object.}
}
\value{
A tibble with option analytics.
}
\description{
Summarise an option
}
