% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_analytics_GQL.R
\name{option_theta_GQL}
\alias{option_theta_GQL}
\title{Option theta}
\usage{
option_theta_GQL(option)
}
\arguments{
\item{option}{QuantLib option object.}
}
\value{
Numeric theta.
}
\description{
Option theta
}
