% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_analytics_GQL.R
\name{option_vega_GQL}
\alias{option_vega_GQL}
\title{Option vega}
\usage{
option_vega_GQL(option)
}
\arguments{
\item{option}{QuantLib option object.}
}
\value{
Numeric vega.
}
\description{
Option vega
}
