% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_helpers_GQH.R
\name{path_tbl_GQH}
\alias{path_tbl_GQH}
\title{Convert path-like simulation output to a tibble}
\usage{
path_tbl_GQH(path)
}
\arguments{
\item{path}{Path object, matrix, data frame, or numeric vector.}
}
\value{
Tibble.
}
\description{
Convert path-like simulation output to a tibble
}
