% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_period_utils_GQL.R
\name{period_chr_GQL}
\alias{period_chr_GQL}
\title{Build and print a QuantLib Period}
\usage{
period_chr_GQL(x)
}
\arguments{
\item{x}{Tenor string such as 1D, 1W, 3M, or 2Y.}
}
\value{
Character scalar.
}
\description{
Build and print a QuantLib Period
}
