% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_period_utils_GQL.R
\name{period_days_GQL}
\alias{period_days_GQL}
\title{Build a QuantLib day Period}
\usage{
period_days_GQL(n)
}
\arguments{
\item{n}{Number of days.}
}
\value{
A QuantLib Period object.
}
\description{
Build a QuantLib day Period
}
