% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_period_utils_GQL.R
\name{period_months_GQL}
\alias{period_months_GQL}
\title{Build a QuantLib month Period}
\usage{
period_months_GQL(n)
}
\arguments{
\item{n}{Number of months.}
}
\value{
A QuantLib Period object.
}
\description{
Build a QuantLib month Period
}
