% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_period_utils_GQL.R
\name{period_weeks_GQL}
\alias{period_weeks_GQL}
\title{Build a QuantLib week Period}
\usage{
period_weeks_GQL(n)
}
\arguments{
\item{n}{Number of weeks.}
}
\value{
A QuantLib Period object.
}
\description{
Build a QuantLib week Period
}
