% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_period_utils_GQL.R
\name{period_years_GQL}
\alias{period_years_GQL}
\title{Build a QuantLib year Period}
\usage{
period_years_GQL(n)
}
\arguments{
\item{n}{Number of years.}
}
\value{
A QuantLib Period object.
}
\description{
Build a QuantLib year Period
}
