% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_constraints_GQL.R
\name{portfolio_constraint_GQL}
\alias{portfolio_constraint_GQL}
\title{Add a Portfolio Constraint}
\usage{
portfolio_constraint_GQL(
  portfolio,
  type = c("weight_sum", "box", "group", "factor_exposure", "position_limit", "turnover",
    "transaction_cost", "diversification", "leverage", "return"),
  ...
)
}
\arguments{
\item{portfolio}{A portfolio specification created by
\code{portfolio_spec_GQL()}.}

\item{type}{Constraint type.}

\item{...}{Constraint parameters.}
}
\value{
Updated portfolio specification.
}
\description{
Add a constraint to a GaussQuant portfolio specification.
}
