% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_objectives_GQL.R
\name{portfolio_objective_GQL}
\alias{portfolio_objective_GQL}
\title{Add a Portfolio Objective}
\usage{
portfolio_objective_GQL(
  portfolio,
  type = c("return", "risk", "risk_budget", "quadratic_utility", "weight_concentration",
    "turnover"),
  measure = NULL,
  ...
)
}
\arguments{
\item{portfolio}{A portfolio specification created by
\code{portfolio_spec_GQL()}.}

\item{type}{Objective type.}

\item{measure}{Optional risk or return measure.}

\item{...}{Additional objective parameters.}
}
\value{
Updated portfolio specification.
}
\description{
Add an objective to a GaussQuant portfolio specification.
}
