% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_spec_GQL.R
\name{portfolio_spec_GQL}
\alias{portfolio_spec_GQL}
\title{Create a Portfolio Specification}
\usage{
portfolio_spec_GQL(assets, weights = NULL)
}
\arguments{
\item{assets}{Character vector of asset names.}

\item{weights}{Optional initial portfolio weights.}
}
\value{
A GaussQuant portfolio specification object.
}
\description{
Create a tidy portfolio specification used by GaussQuant portfolio
optimisation functions.
}
