% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exotic_derivatice_FX_IR.R
\name{prdc_cashflow_table_GQL}
\alias{prdc_cashflow_table_GQL}
\title{Build a PRDC expected cash-flow table}
\usage{
prdc_cashflow_table_GQL(
  expected_coupon_tbl,
  valuation_date,
  coupon_schedule,
  day_counter,
  discount_curve,
  notional,
  redemption_amount = notional
)
}
\arguments{
\item{expected_coupon_tbl}{Output from \code{prdc_expected_coupon_GQL()}.}

\item{valuation_date}{ISO date, R Date, or QuantLib Date.}

\item{coupon_schedule}{QuantLib Schedule.}

\item{day_counter}{QuantLib day counter.}

\item{discount_curve}{QuantLib discount curve or handle.}

\item{notional}{Product notional.}

\item{redemption_amount}{Redemption amount paid on the final coupon date.}
}
\value{
Tibble containing coupon, redemption, and present values.
}
\description{
Build a PRDC expected cash-flow table
}
