% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exotic_derivatice_FX_IR.R
\name{prdc_coupon_GQL}
\alias{prdc_coupon_GQL}
\title{Calculate a PRDC coupon rate}
\usage{
prdc_coupon_GQL(
  fx_rate,
  fx_base,
  leverage,
  subtraction,
  floor_rate = 0,
  cap_rate = Inf
)
}
\arguments{
\item{fx_rate}{Numeric FX rate vector or matrix.}

\item{fx_base}{Base FX rate in the coupon formula.}

\item{leverage}{FX leverage coefficient.}

\item{subtraction}{Fixed subtraction rate.}

\item{floor_rate}{Coupon floor.}

\item{cap_rate}{Coupon cap.}
}
\value{
Coupon rates with the same dimensions as \code{fx_rate}.
}
\description{
Calculate a PRDC coupon rate
}
