% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exotic_derivatice_FX_IR.R
\name{prdc_path_matrix_GQL}
\alias{prdc_path_matrix_GQL}
\title{Generate a PRDC FX path matrix}
\usage{
prdc_path_matrix_GQL(path_generator, n_paths, seed = 1L)
}
\arguments{
\item{path_generator}{Object returned by \code{fx_path_generator_GQL()}.}

\item{n_paths}{Number of Monte Carlo paths.}

\item{seed}{Random seed.}
}
\value{
Numeric matrix with coupon dates in rows and paths in columns.
}
\description{
Generate a PRDC FX path matrix
}
