% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{push_calibration_helpers_GQL}
\alias{push_calibration_helpers_GQL}
\title{Push calibration helpers into a QuantLib vector}
\usage{
push_calibration_helpers_GQL(helpers)
}
\arguments{
\item{helpers}{List of QuantLib calibration helpers.}
}
\value{
QuantLib CalibrationHelperVector.
}
\description{
Push calibration helpers into a QuantLib vector
}
