% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_GQL.R
\name{quote_handle_GQL}
\alias{quote_handle_GQL}
\title{Create QuantLib QuoteHandle}
\usage{
quote_handle_GQL(x)
}
\arguments{
\item{x}{Numeric quote.}
}
\value{
QuantLib QuoteHandle.
}
\description{
Create QuantLib QuoteHandle
}
