% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{replace_metric_value_GQH}
\alias{replace_metric_value_GQH}
\title{Replace or append a metric value}
\usage{
replace_metric_value_GQH(tbl, metric_name, new_value)
}
\arguments{
\item{tbl}{Metric table.}

\item{metric_name}{Metric name.}

\item{new_value}{Replacement value.}
}
\value{
Updated metric table.
}
\description{
Replace or append a metric value
}
