% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{rmse_GQH}
\alias{rmse_GQH}
\title{Calculate root mean squared error}
\usage{
rmse_GQH(actual, fitted)
}
\arguments{
\item{actual}{Actual values.}

\item{fitted}{Fitted values.}
}
\value{
Numeric RMSE.
}
\description{
Calculate root mean squared error
}
