% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{safe_iso_GQH}
\alias{safe_iso_GQH}
\title{Safely convert a QuantLib date to ISO}
\usage{
safe_iso_GQH(x)
}
\arguments{
\item{x}{QuantLib Date object or NULL.}
}
\value{
ISO date string or NA.
}
\description{
Safely convert a QuantLib date to ISO
}
