% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{safe_sabr_vol_GQL}
\alias{safe_sabr_vol_GQL}
\title{Safely calculate Black SABR volatility}
\usage{
safe_sabr_vol_GQL(strike, forward, maturity, alpha, beta, volvol, rho)
}
\arguments{
\item{strike}{Strike.}

\item{forward}{Forward.}

\item{maturity}{Maturity.}

\item{alpha}{SABR alpha.}

\item{beta}{SABR beta.}

\item{volvol}{SABR volatility of volatility.}

\item{rho}{SABR correlation.}
}
\value{
Numeric volatility or NA.
}
\description{
Safely calculate Black SABR volatility
}
