% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_period_utils_GQL.R
\name{schedule_date_at_GQL}
\alias{schedule_date_at_GQL}
\title{Extract a date from a QuantLib schedule}
\usage{
schedule_date_at_GQL(schedule, i_one_based)
}
\arguments{
\item{schedule}{QuantLib Schedule object.}

\item{i_one_based}{One-based date index.}
}
\value{
A QuantLib Date object, or NULL if extraction fails.
}
\description{
Extract a date from a QuantLib schedule
}
