% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_period_utils_GQL.R
\name{schedule_date_at_local_GQL}
\alias{schedule_date_at_local_GQL}
\title{Extract a schedule date as local R Date}
\usage{
schedule_date_at_local_GQL(schedule, i_one_based)
}
\arguments{
\item{schedule}{QuantLib Schedule object.}

\item{i_one_based}{One-based date index.}
}
\value{
An R Date, or NA if extraction fails.
}
\description{
Extract a schedule date as local R Date
}
