% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{schedule_date_vector_GQL}
\alias{schedule_date_vector_GQL}
\title{Build an R Date vector from a QuantLib schedule}
\usage{
schedule_date_vector_GQL(schedule)
}
\arguments{
\item{schedule}{QuantLib Schedule object.}
}
\value{
An R Date vector.
}
\description{
Build an R Date vector from a QuantLib schedule
}
