% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_period_utils_GQL.R
\name{schedule_dates_GQL}
\alias{schedule_dates_GQL}
\title{Build a table of QuantLib schedule dates}
\usage{
schedule_dates_GQL(schedule)
}
\arguments{
\item{schedule}{QuantLib Schedule object.}
}
\value{
A tibble with schedule_date.
}
\description{
Build a table of QuantLib schedule dates
}
