% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{schedule_table_GQL}
\alias{schedule_table_GQL}
\title{Build a schedule table with period numbers}
\usage{
schedule_table_GQL(schedule)
}
\arguments{
\item{schedule}{QuantLib Schedule object.}
}
\value{
Tibble with period and schedule_date.
}
\description{
Build a schedule table with period numbers
}
