% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics_GQL.R
\name{set_bond_pricing_engine_GQL}
\alias{set_bond_pricing_engine_GQL}
\title{Set Bond Pricing Engine}
\usage{
set_bond_pricing_engine_GQL(bond, curve)
}
\arguments{
\item{bond}{QuantLib bond object.}

\item{curve}{QuantLib yield term structure.}
}
\value{
Bond with pricing engine.
}
\description{
Set Bond Pricing Engine
}
