% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_period_utils_GQL.R
\name{set_eval_date_GQL}
\alias{set_eval_date_GQL}
\title{Set QuantLib evaluation date}
\usage{
set_eval_date_GQL(eval_date)
}
\arguments{
\item{eval_date}{ISO date string, R Date, or QuantLib Date.}
}
\value{
The QuantLib Date invisibly.
}
\description{
Set QuantLib evaluation date
}
