% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{shifted_normal_vol_hagan_GQL}
\alias{shifted_normal_vol_hagan_GQL}
\title{Calculate shifted Hagan normal SABR volatility}
\usage{
shifted_normal_vol_hagan_GQL(
  strike,
  forward,
  maturity,
  beta,
  alpha,
  volvol,
  rho,
  shift = 0.025
)
}
\arguments{
\item{strike}{Strike.}

\item{forward}{Forward.}

\item{maturity}{Maturity.}

\item{beta}{SABR beta.}

\item{alpha}{SABR alpha.}

\item{volvol}{SABR volatility of volatility.}

\item{rho}{SABR correlation.}

\item{shift}{Shift applied to strike and forward.}
}
\value{
Numeric normal volatility or NA.
}
\description{
Calculate shifted Hagan normal SABR volatility
}
