% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_bootstrap_sofr_GQL.R
\name{sofr_curve_nodes_GQL}
\alias{sofr_curve_nodes_GQL}
\title{Extract SOFR curve nodes safely}
\usage{
sofr_curve_nodes_GQL(curve)
}
\arguments{
\item{curve}{QuantLib SOFR curve object.}
}
\value{
A tibble containing node dates and continuously compounded forward
rates.
}
\description{
Avoid direct use of \code{curve$nodes()}, which is unstable in some QuantLib
SWIG environments.
}
