% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_bootstrap_sofr_GQL.R
\name{sofr_forward_rate_GQL}
\alias{sofr_forward_rate_GQL}
\title{Calculate a SOFR forward rate}
\usage{
sofr_forward_rate_GQL(
  curve,
  start_date,
  end_date,
  day_counter = QuantLib::Actual360(),
  compounding = QuantLib::Compounding_Continuous_get()
)
}
\arguments{
\item{curve}{QuantLib SOFR curve object.}

\item{start_date}{Start date in ISO format.}

\item{end_date}{End date in ISO format.}

\item{day_counter}{QuantLib day-counter object.}

\item{compounding}{QuantLib compounding convention.}
}
\value{
A numeric forward rate.
}
\description{
Setting \code{start_date} and \code{end_date} to the same date returns the
instantaneous forward rate implied by the curve.
}
