% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_bootstrap_sofr_GQL.R
\name{sofr_future_rate_helper_GQL}
\alias{sofr_future_rate_helper_GQL}
\title{Create a SOFR futures rate helper}
\usage{
sofr_future_rate_helper_GQL(
  price,
  reference_month,
  reference_year,
  reference_frequency = c("Quarterly", "Monthly")
)
}
\arguments{
\item{price}{SOFR futures price.}

\item{reference_month}{Contract reference month, such as \code{"March"}.}

\item{reference_year}{Contract reference year.}

\item{reference_frequency}{\code{"Quarterly"} or \code{"Monthly"}.}
}
\value{
A QuantLib SOFR futures rate-helper object.
}
\description{
Create a SOFR futures rate helper
}
