% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_bootstrap_sofr_GQL.R
\name{sofr_futures_price_GQL}
\alias{sofr_futures_price_GQL}
\title{Convert a SOFR futures rate quote to price}
\usage{
sofr_futures_price_GQL(rate_pct)
}
\arguments{
\item{rate_pct}{Interest-rate quote in percent.}
}
\value{
A numeric SOFR futures price.
}
\description{
Convert a SOFR futures rate quote to price
}
