% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_bootstrap_sofr_GQL.R
\name{sofr_GQL}
\alias{sofr_GQL}
\title{Create a SOFR overnight index}
\usage{
sofr_GQL(forwarding_curve_handle = NULL)
}
\arguments{
\item{forwarding_curve_handle}{Optional QuantLib yield-term-structure
handle.}
}
\value{
A QuantLib SOFR overnight-index object.
}
\description{
Create a SOFR overnight index
}
