% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_GQL.R
\name{str_date_ISO_GQL}
\alias{str_date_ISO_GQL}
\title{Convert a QuantLib Date to an ISO date string}
\usage{
str_date_ISO_GQL(date_obj)
}
\arguments{
\item{date_obj}{A QuantLib Date object.}
}
\value{
A character scalar in YYYY-MM-DD format.
}
\description{
Convert a QuantLib Date to an ISO date string
}
