% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{swap_annuity_from_schedule_GQL}
\alias{swap_annuity_from_schedule_GQL}
\title{Calculate swap annuity from a schedule}
\usage{
swap_annuity_from_schedule_GQL(schedule, curve, day_counter)
}
\arguments{
\item{schedule}{QuantLib Schedule.}

\item{curve}{QuantLib yield curve.}

\item{day_counter}{QuantLib day counter.}
}
\value{
Numeric annuity.
}
\description{
Calculate swap annuity from a schedule
}
