% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics_GQL.R
\name{swap_cashflow_schedule_GQL}
\alias{swap_cashflow_schedule_GQL}
\title{Build cashflow schedule table from swap legs}
\usage{
swap_cashflow_schedule_GQL(swap)
}
\arguments{
\item{swap}{A QuantLib VanillaSwap-like object.}
}
\value{
A tibble containing fixed and floating leg cashflows.
}
\description{
Build cashflow schedule table from swap legs
}
