% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics_GQL.R
\name{swap_fair_rate_GQL}
\alias{swap_fair_rate_GQL}
\title{Swap fair fixed rate}
\usage{
swap_fair_rate_GQL(swap)
}
\arguments{
\item{swap}{A QuantLib VanillaSwap-like object.}
}
\value{
Fair fixed rate. Returns NA if unavailable.
}
\description{
Swap fair fixed rate
}
