% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics_GQL.R
\name{swap_fair_spread_GQL}
\alias{swap_fair_spread_GQL}
\title{Swap fair spread}
\usage{
swap_fair_spread_GQL(swap)
}
\arguments{
\item{swap}{A QuantLib VanillaSwap-like object.}
}
\value{
Fair spread. Returns NA if unavailable.
}
\description{
Swap fair spread
}
