% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics_GQL.R
\name{swap_fixed_leg_npv_GQL}
\alias{swap_fixed_leg_npv_GQL}
\title{Swap fixed leg NPV}
\usage{
swap_fixed_leg_npv_GQL(swap)
}
\arguments{
\item{swap}{A QuantLib VanillaSwap-like object.}
}
\value{
Fixed leg NPV.
}
\description{
Swap fixed leg NPV
}
