% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics_GQL.R
\name{swap_fixed_leg_table_GQL}
\alias{swap_fixed_leg_table_GQL}
\title{Swap fixed leg cashflow table}
\usage{
swap_fixed_leg_table_GQL(swap)
}
\arguments{
\item{swap}{A QuantLib VanillaSwap-like object.}
}
\value{
A tibble of fixed leg cashflows.
}
\description{
Swap fixed leg cashflow table
}
