% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics_GQL.R
\name{swap_floating_leg_npv_GQL}
\alias{swap_floating_leg_npv_GQL}
\title{Swap floating leg NPV}
\usage{
swap_floating_leg_npv_GQL(swap)
}
\arguments{
\item{swap}{A QuantLib VanillaSwap-like object.}
}
\value{
Floating leg NPV.
}
\description{
Swap floating leg NPV
}
