% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics_GQL.R
\name{swap_floating_leg_table_GQL}
\alias{swap_floating_leg_table_GQL}
\title{Swap floating leg cashflow table}
\usage{
swap_floating_leg_table_GQL(swap)
}
\arguments{
\item{swap}{A QuantLib VanillaSwap-like object.}
}
\value{
A tibble of floating leg cashflows.
}
\description{
Swap floating leg cashflow table
}
