% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics_GQL.R
\name{swap_leg_npv_GQL}
\alias{swap_leg_npv_GQL}
\title{Swap leg NPV}
\usage{
swap_leg_npv_GQL(swap, leg_no)
}
\arguments{
\item{swap}{A QuantLib Swap object.}

\item{leg_no}{Leg number. QuantLib uses 0-based leg indexing.}
}
\value{
Swap leg NPV.
}
\description{
Swap leg NPV
}
