% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics_GQL.R
\name{swap_summary_GQL}
\alias{swap_summary_GQL}
\title{Swap summary}
\usage{
swap_summary_GQL(swap)
}
\arguments{
\item{swap}{A QuantLib VanillaSwap-like object.}
}
\value{
A tibble summarising swap valuation measures.
}
\description{
Swap summary
}
