% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option_GQL.R
\name{swaption_annuity_GQL}
\alias{swaption_annuity_GQL}
\title{Calculate swaption annuity}
\usage{
swaption_annuity_GQL(swaption)
}
\arguments{
\item{swaption}{QuantLib Swaption object.}
}
\value{
Numeric annuity if available.
}
\description{
Calculate swaption annuity
}
