% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option_GQL.R
\name{swaption_npv_GQL}
\alias{swaption_npv_GQL}
\title{Calculate swaption NPV}
\usage{
swaption_npv_GQL(swaption)
}
\arguments{
\item{swaption}{QuantLib Swaption object.}
}
\value{
Numeric NPV.
}
\description{
Calculate swaption NPV
}
