% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option_GQL.R
\name{swaption_summary_GQL}
\alias{swaption_summary_GQL}
\title{Summarise a swaption}
\usage{
swaption_summary_GQL(swaption)
}
\arguments{
\item{swaption}{QuantLib Swaption object.}
}
\value{
A tibble with swaption analytics.
}
\description{
Summarise a swaption
}
