% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option_GQL.R
\name{swaption_vega_GQL}
\alias{swaption_vega_GQL}
\title{Calculate swaption vega}
\usage{
swaption_vega_GQL(swaption)
}
\arguments{
\item{swaption}{QuantLib Swaption object.}
}
\value{
Numeric vega if available.
}
\description{
Calculate swaption vega
}
