% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_bootstrap_tonar_GQL.R
\name{tonar_curve_nodes_GQL}
\alias{tonar_curve_nodes_GQL}
\title{Extract TONAR curve nodes safely}
\usage{
tonar_curve_nodes_GQL(curve)
}
\arguments{
\item{curve}{QuantLib TONAR curve object.}
}
\value{
A tibble containing node dates and continuously compounded forward
rates.
}
\description{
Avoid direct use of \code{curve$nodes()}, which is unstable in some QuantLib
SWIG environments.
}
