% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_bootstrap_tonar_GQL.R
\name{tonar_GQL}
\alias{tonar_GQL}
\title{Create a TONAR overnight index}
\usage{
tonar_GQL(forwarding_curve_handle = NULL)
}
\arguments{
\item{forwarding_curve_handle}{Optional QuantLib yield-term-structure
handle.}
}
\value{
A QuantLib TONAR overnight-index object.
}
\description{
Create a TONAR overnight index
}
