% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat_quantive_riemann_GQH.R
\name{trade_ois_swap_py_GQH}
\alias{trade_ois_swap_py_GQH}
\title{Python-style OIS swap trade wrapper}
\usage{
trade_ois_swap_py_GQH(
  curve_env,
  effective = NULL,
  maturity,
  notional = 1e+06,
  fixed_rate,
  float_spread = 0,
  pay_receive = c("pay", "receive"),
  fixed_schedule_tenor = NULL,
  floating_schedule_tenor = NULL,
  pay_lag = NULL,
  fixed_bdc = "ModifiedFollowing",
  date_rule = "Backward",
  eom = FALSE,
  verbose = TRUE
)
}
\description{
Python-style OIS swap trade wrapper
}
