% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{use_quantlib_GQH}
\alias{use_quantlib_GQH}
\title{Add a local QuantLib library path}
\usage{
use_quantlib_GQH(
  path = "C:/AnalyticFin/Projects/QuantLib_tidy/library/windows"
)
}
\arguments{
\item{path}{Local library path containing the SWIG QuantLib package.}
}
\value{
The updated library search paths, invisibly.
}
\description{
Add a local QuantLib library path
}
