% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{vasicek_b_GQL}
\alias{vasicek_b_GQL}
\title{Calculate Vasicek B}
\usage{
vasicek_b_GQL(time, maturity, mean_reversion)
}
\arguments{
\item{time}{Start time.}

\item{maturity}{Maturity.}

\item{mean_reversion}{Mean reversion.}
}
\value{
Numeric B.
}
\description{
Calculate Vasicek B
}
