% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_builder_GQL.R
\name{yield_curve_handle_GQL}
\alias{yield_curve_handle_GQL}
\title{Build a QuantLib yield curve handle}
\usage{
yield_curve_handle_GQL(curve)
}
\arguments{
\item{curve}{A QuantLib yield term structure object.}
}
\value{
A QuantLib YieldTermStructureHandle object.
}
\description{
Build a QuantLib yield curve handle
}
