% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics_GQL.R
\name{zero_coupon_bond_GQL}
\alias{zero_coupon_bond_GQL}
\title{Create a zero coupon bond}
\usage{
zero_coupon_bond_GQL(
  discount_curve,
  maturity = "2013-08-15",
  issue = "2008-08-15",
  face = 100
)
}
\description{
Create a zero coupon bond
}
