% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_GQL.R
\name{date_GQL}
\alias{date_GQL}
\title{Convert ISO date to QuantLib Date}
\usage{
date_GQL(x)
}
\arguments{
\item{x}{Date or character scalar.}
}
\value{
QuantLib Date object.
}
\description{
Convert ISO date to QuantLib Date
}
