% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{mid_date_GQH}
\alias{mid_date_GQH}
\title{Calculate a midpoint date}
\usage{
mid_date_GQH(first_date, second_date)
}
\arguments{
\item{first_date}{First QuantLib date.}

\item{second_date}{Second QuantLib date.}
}
\value{
QuantLib midpoint date.
}
\description{
Calculate a midpoint date
}
