% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_optimize_GQL.R
\name{portfolio_optimize_GQL}
\alias{portfolio_optimize_GQL}
\title{Optimize a Portfolio}
\usage{
portfolio_optimize_GQL(
  tbl_returns,
  portfolio,
  date_col = "date",
  asset_col = "asset",
  return_col = "return",
  method = c("minimum_variance", "mean_variance", "maximum_return"),
  solver = c("ROI", "optim"),
  risk_aversion = 1,
  control = list()
)
}
\arguments{
\item{tbl_returns}{A tibble containing asset returns.}

\item{portfolio}{A portfolio specification created by
\code{portfolio_spec_GQL()}.}

\item{date_col}{Name of the date column.}

\item{asset_col}{Name of the asset column.}

\item{return_col}{Name of the return column.}

\item{method}{Portfolio optimization method.}

\item{solver}{Optimization solver.}

\item{risk_aversion}{Risk-aversion parameter used by mean-variance
optimization.}

\item{control}{Optional solver control list.}
}
\value{
An object of class \code{portfolio_optimization_GQL}.
}
\description{
Optimize a GaussQuant portfolio specification using asset return data.
}
