% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GQH.R
\name{schedule_date_vector_GQL}
\alias{schedule_date_vector_GQL}
\title{Build a list of dates from a QuantLib schedule}
\usage{
schedule_date_vector_GQL(schedule)
}
\arguments{
\item{schedule}{QuantLib Schedule object.}
}
\value{
List of QuantLib Date objects.
}
\description{
Build a list of dates from a QuantLib schedule
}
